pyproject.toml
[tool.poetry]
name = "emerg_shutdown"
version = "0.1.0"
description = "Learning application based on a real world requirement"
authors = ["dpnetca <git@dpnet.ca>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.63.0"
uvicorn = "^0.13.2"
aiofiles = "^0.6.0"
jinja2 = "^2.11.2"
python-multipart = "^0.0.5"
pymongo = "^3.11.2"
mongoengine = "^0.21.0"
python-dotenv = "^0.15.0"
zeep = {extras = ["async"], version = "^4.0.0"}
python-simplexml = "^0.1.5"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
bandit = "^1.7.0"
black = "^20.8b1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
